pyproject.toml
[tool.poetry]
name = "g-sheets-tools"
version = "0.1.5"
description = ""
authors = ["Andy Friedman <afriedman412@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7.1"
cachetools = "4.1.1"
chardet = "^4.0.0"
certifi = "^2021.10.8"
google-auth = "^2.3.3"
google-auth-httplib2 = "^0.1.0"
google-api-client = "^3.14.159265359"
google-api-python-client = "^2.33.0"
google-api-core = "^2.3.2"
google-auth-oauthlib = "^0.4.6"
googleapis-common-protos = "^1.54.0"
httplib2 = "^0.20.2"
idna = "^3.3"
numpy = "^1.19.2"
oauthlib = "^3.1.1"
pandas = "^1.3.5"
protobuf = "^3.19.1"
pyasn1 = "^0.4.8"
pyasn1-modules = "^0.2.8"
python-dateutil = "^2.8.2"
pytz = "^2021.3"
requests = "^2.26.0"
requests-oauthlib = "^1.3.0"
rsa = "^4.8"
six = "^1.16.0"
uritemplate = "^4.1.1"
urllib3 = "^1.26.7"
pytest = "^6.2.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
